+++ mapentity/registry.py
@@ -164,7 +164,9 @@ class MapEntityOptions:
             mapentity_models.ENTITY_MAPIMAGE: r"^image/{modelname}-(?P<pk>\d+).png$",
             mapentity_models.ENTITY_CREATE: r"^{modelname}/add/$",
             mapentity_models.ENTITY_UPDATE: r"^{modelname}/edit/(?P<pk>\d+)/$",
+            mapentity_models.ENTITY_MULTI_UPDATE: r"^{modelname}/multi_update/$",
             mapentity_models.ENTITY_DELETE: r"^{modelname}/delete/(?P<pk>\d+)/$",
+            mapentity_models.ENTITY_MULTI_DELETE: r"^{modelname}/multi_delete/$",
             mapentity_models.ENTITY_MARKUP: r"^{modelname}/markup/(?P<pk>\d+)/$",
         }
         if app_settings["MAPENTITY_WEASYPRINT"]: