+++ test_project/test_app/models.py
@@ -1,3 +1,5 @@
+from django.contrib.contenttypes.fields import GenericForeignKey
+from django.contrib.contenttypes.models import ContentType
 from django.contrib.gis.db import models
 from django.contrib.gis.geos import GEOSGeometry
 from django.utils.translation import gettext_lazy as _
@@ -138,6 +140,27 @@ class Sector(MapEntityMixin, models.Model):
     code = models.CharField(primary_key=True, max_length=6)
     name = models.CharField(max_length=100)
     skip_attachments = True
+
+    def __str__(self):
+        return self.name
+
+
+class GeoPoint(MapEntityMixin, models.Model):
+    public = models.BooleanField(default=False)
+    geom = models.PointField(null=True, default=None)
+    located_in = models.ForeignKey(
+        City, on_delete=models.SET_NULL, null=True, blank=True
+    )
+    road = models.ForeignKey(Road, on_delete=models.CASCADE)
+    name = models.CharField(max_length=255)
+    dummy_model = models.OneToOneField(DummyModel, null=True, on_delete=models.CASCADE)
+    internal_reference = models.CharField(max_length=20, editable=False)
+    content_type = models.ForeignKey(
+        ContentType, blank=True, null=True, on_delete=models.SET_NULL
+    )
+    object_id = models.PositiveIntegerField(blank=True, null=True)
+    related_object = GenericForeignKey("content_type", "object_id")
+    tags = models.ManyToManyField(Tag, blank=True)
 
     def __str__(self):
         return self.name