+++ test_project/test_app/translation.py
@@ -1,6 +1,6 @@
 from modeltranslation.translator import TranslationOptions, translator
 
-from test_project.test_app.models import DummyModel, Road
+from test_project.test_app.models import DummyModel, GeoPoint, Road
 
 
 class DummyModelTO(TranslationOptions):
@@ -11,5 +11,10 @@ class RoadModelTO(TranslationOptions):
     fields = ("name",)
 
 
+class GeoPointModelTO(TranslationOptions):
+    fields = ("name", "public")
+
+
 translator.register(DummyModel, DummyModelTO)
 translator.register(Road, RoadModelTO)
+translator.register(GeoPoint, GeoPointModelTO)