+++ test_project/test_app/urls.py
@@ -1,5 +1,5 @@
 from mapentity.registry import registry
-from test_project.test_app.models import City, DummyModel, MushroomSpot, Road
+from test_project.test_app.models import City, DummyModel, GeoPoint, MushroomSpot, Road
 
 app_name = "test_app"
 
@@ -7,3 +7,4 @@ urlpatterns = registry.register(DummyModel)
 urlpatterns += registry.register(MushroomSpot)
 urlpatterns += registry.register(Road)
 urlpatterns += registry.register(City)
+urlpatterns += registry.register(GeoPoint)